+++ packages/eyes-protractor/src/spec-driver.js
@@ -242,12 +242,15 @@ async function build(env) {
 
   const desiredCapabilities = {browserName: browser, ...capabilities}
   if (configurable) {
-    const browserOptionsName = browserOptionsNames[browser]
+    const browserOptionsName = browserOptionsNames[browser || desiredCapabilities.browserName]
     if (browserOptionsName) {
-      desiredCapabilities[browserOptionsName] = {
-        args: headless ? args.concat('headless') : args,
-        debuggerAddress: attach === true ? 'localhost:9222' : attach,
+      const browserOptions = desiredCapabilities[browserOptionsName] || {}
+      browserOptions.args = [...(browserOptions.args || []), ...args]
+      if (headless) browserOptions.args.push('headless')
+      if (attach) {
+        browserOptions.debuggerAddress = attach === true ? 'localhost:9222' : attach
       }
+      desiredCapabilities[browserOptionsName] = browserOptions
     }
   }
   const builder = new Builder().withCapabilities(desiredCapabilities)