+++ packages/eyes-webdriverio-5/.mocharc.cvg.js
@@ -15,8 +15,8 @@ const tags = {
 const protocol = process.env.APPLITOOLS_WDIO_PROTOCOL in tags ? process.env.APPLITOOLS_WDIO_PROTOCOL : 'wd'
 module.exports = {
   spec: [
-    './test/coverage/generic/*.spec.js',
-    '../sdk-shared/coverage-tests/custom/**.spec.js',
+    './test/generic/*.spec.js',
+    '../sdk-shared/coverage-tests/custom/**/*.spec.js',
   ],
   parallel: true,
   jobs: 15,