+++ packages/sdk-shared/coverage-tests/custom/TestEyesDifferentRunners/TestEyesDifferentRunnersSelenium.spec.js
@@ -37,6 +37,6 @@ describe('TestEyesDifferentRunners Selenium', () => {
     ['https://wikipedia.org', MatchLevel.Strict],
   ]
   cases.forEach(testData => {
-    it(testData[0], testCase(...testData))
+    it('TestEyesDifferentRunners', testCase(...testData))
   })
 })