+++ packages/sdk-shared/coverage-tests/custom/TestEyesDifferentRunners/TestEyesDifferentRunnersVG2.spec.js
@@ -4,7 +4,7 @@ const path = require('path')
 const {getEyes} = require('../../../src/test-setup')
 const spec = require(path.resolve(cwd, 'src/spec-driver'))
 const {MatchLevel} = require(cwd)
-const {testSetup, getCheckSettings} = require('./EyesDifferentRunners')
+const {testSetup, getCheckSettings, validateVG2} = require('./EyesDifferentRunners')
 
 describe('TestEyesDifferentRunners VG2', () => {
   afterEach(async function() {
@@ -14,7 +14,7 @@ describe('TestEyesDifferentRunners VG2', () => {
 
   beforeEach(async function() {
     ;[this.webDriver, this.destroyDriver] = await spec.build({browser: 'chrome'})
-    this.eyes = await getEyes({isVisualGrid: true})
+    this.eyes = await getEyes({vg: true})
     this.eyes.setSaveNewTests(false)
     await this.eyes.open(this.webDriver, 'Top Sites', `Top Sites - ${this.currentTest.title}`, {
       width: 800,
@@ -22,11 +22,9 @@ describe('TestEyesDifferentRunners VG2', () => {
     })
   })
 
-  let testCase = testSetup(getCheckSettings, async () =>
-    console.log('Need merge of the runners updates to retrieve the test results for assertions'),
-  )
+  let testCase = testSetup(getCheckSettings, validateVG2)
   let cases = [['https://amazon.com', MatchLevel.Layout]]
   cases.forEach(testData => {
-    it(testData[0], testCase(...testData))
+    it('TestEyesDifferentRunners', testCase(...testData))
   })
 })