+++ packages/sdk-shared/coverage-tests/custom/TestEyesDifferentRunners/TestEyesDifferentRunnersVGHook.spec.js
@@ -14,7 +14,7 @@ describe('TestEyesDifferentRunners VG with hooks', () => {
 
   beforeEach(async function() {
     ;[this.webDriver, this.destroyDriver] = await spec.build({browser: 'chrome'})
-    this.eyes = await getEyes({isVisualGrid: true})
+    this.eyes = await getEyes({vg: true})
     let conf = this.eyes.getConfiguration()
     conf.setTestName(`Top Sites - ${this.currentTest.title}`)
     conf.setAppName(`Top Sites`)
@@ -34,6 +34,6 @@ describe('TestEyesDifferentRunners VG with hooks', () => {
   ]
 
   cases.forEach(testData => {
-    it(testData[0], testCase(...testData))
+    it('TestEyesDifferentRunners', testCase(...testData))
   })
 })