+++ packages/sdk-shared/coverage-tests/custom/TestMobileDevices/TestIOSSafariStitch/Portrait/TestiPhone7Simulator10.spec.js
@@ -1,0 +1,18 @@
+'use strict'
+const {getMobileEmulation, testMobileDevices, iPhoneAgent} = require('../../TestMobileDevices')
+let device = {
+  mobileEmulation: getMobileEmulation(iPhoneAgent, 375, 559, 2),
+  name: 'iPhone 7 Simulator 10.3',
+  orientation: 'Portrait',
+}
+describe.skip(`${device.name} Portrait`, () => {
+  describe(`mobile`, () => {
+    it('TestIOSSafariStitch', testMobileDevices(device, 'mobile'))
+  })
+  describe(`desktop`, () => {
+    it('TestIOSSafariStitch', testMobileDevices(device, 'desktop'))
+  })
+  describe(`scrolled_mobile`, () => {
+    it('TestIOSSafariStitch', testMobileDevices(device, 'scrolled_mobile'))
+  })
+})