+++ packages/sdk-shared/coverage-tests/custom/TestMobileDevices/TestIOSSafariStitch/Portrait/TestiPhoneXR12.spec.js
@@ -1,0 +1,18 @@
+'use strict'
+const {getMobileEmulation, testMobileDevices, iPhoneAgent} = require('../../TestMobileDevices')
+let device = {
+  mobileEmulation: getMobileEmulation(iPhoneAgent, 414, 719, 2),
+  name: 'iPhone XR Simulator 12.2',
+  orientation: 'Portrait',
+}
+describe(`${device.name} Portrait`, () => {
+  describe(`mobile`, () => {
+    it('TestIOSSafariStitch', testMobileDevices(device, 'mobile'))
+  })
+  describe(`desktop`, () => {
+    it.skip('TestIOSSafariStitch', testMobileDevices(device, 'desktop'))
+  })
+  describe(`scrolled_mobile`, () => {
+    it('TestIOSSafariStitch', testMobileDevices(device, 'scrolled_mobile'))
+  })
+})