+++ packages/sdk-shared/coverage-tests/custom/TestVisualGridRefererHeader.spec.js
@@ -39,7 +39,7 @@ describe('TestVisualGridRefererHeader', () => {
   it('send referer header', async () => {
     const url = adjustUrlToDocker('http://localhost:5555/cors.html')
     await spec.visit(driver, url)
-    const eyes = getEyes({isVisualGrid: true})
+    const eyes = getEyes({vg: true})
     await eyes.open(driver, 'VgFetch', ' VgFetch referer', {width: 800, height: 600})
     await eyes.check('referer', Target.window())
     await eyes.close()