+++ packages/eyes-protractor/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - deprecate `saveDebugData`
+- updated to @applitools/eyes-sdk-core@12.4.4 (from 12.4.2)
+- updated to @applitools/visual-grid-client@15.0.12 (from 15.0.10)
 
 ## 1.9.1 - 2020/10/15